refactor: extract DWARF entry value lowering
Move DW_OP_entry_value recovery and call-site fallback handling into
entry_value.rs. The main lowering path now delegates entry-value decisions
instead of carrying the full caller recovery chain inline.

Keep existing full-resolution and soft fallback behavior unchanged while
moving the entry-value specific tests with the implementation.
